# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AromaType API
  version: 0.1.0
  description: API contract for the AromaType backend MVP.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Backend health check
      operationId: getHealth
      tags:
        - System
      responses:
        "200":
          description: Backend is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                service: aromatype-backend
                version: 0.1.0
  /api/auth/register:
    post:
      summary: Register user
      operationId: registerUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: user@example.com
              password: password123
      responses:
        "201":
          description: User registered and JWT issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      summary: Login user
      operationId: loginUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: user@example.com
              password: password123
      responses:
        "200":
          description: User authenticated and JWT issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/questions:
    get:
      summary: Get questionnaire questions
      operationId: getQuestions
      tags:
        - Questionnaire
      responses:
        "200":
          description: Active questionnaire questions with answer options.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /api/recommendations:
    post:
      summary: Build fragrance recommendations
      description: Builds a basic rule-based recommendation without LLM.
      operationId: buildRecommendations
      tags:
        - Recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRequest"
            example:
              answerOptionIds:
                - q1_a1
                - q2_a1
                - q3_a1
                - q4_a2
                - q5_a2
                - q6_a1
                - q7_a2
      responses:
        "200":
          description: User aroma profile and recommended fragrances.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/fragrances/{id}:
    get:
      summary: Get fragrance product card
      operationId: getFragrance
      tags:
        - Fragrances
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: fresh-office
      responses:
        "200":
          description: Fragrance product card with notes and usage metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fragrance"
        "404":
          description: Fragrance not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/fragrances:
    post:
      summary: Create fragrance as admin
      operationId: createFragrance
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFragranceRequest"
      responses:
        "201":
          description: Fragrance created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fragrance"
        "400":
          description: Invalid fragrance data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/uploads/fragrance-photo:
    post:
      summary: Upload fragrance photo as admin
      operationId: uploadFragrancePhoto
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        "201":
          description: Photo uploaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Invalid upload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
        - version
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: aromatype-backend
        version:
          type: string
          example: 0.1.0
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 6
          example: password123
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: password123
    AuthResponse:
      type: object
      required:
        - accessToken
        - user
      properties:
        accessToken:
          type: string
          description: JWT access token.
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - user
            - admin
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: invalid_input
        message:
          type: string
          example: Request validation failed.
    Question:
      type: object
      required:
        - id
        - text
        - type
        - sortOrder
        - options
      properties:
        id:
          type: string
          example: q1
        text:
          type: string
          example: Если описать вас в трёх словах через аромат, что это будет?
        type:
          type: string
          enum:
            - single_choice
            - multiple_choice
          example: single_choice
        sortOrder:
          type: integer
          example: 1
        options:
          type: array
          items:
            $ref: "#/components/schemas/AnswerOption"
    AnswerOption:
      type: object
      required:
        - id
        - questionId
        - text
        - value
        - sortOrder
      properties:
        id:
          type: string
          example: q1_a1
        questionId:
          type: string
          example: q1
        text:
          type: string
          example: Спокойный и уверенный
        value:
          type: string
          example: calm_confident
        sortOrder:
          type: integer
          example: 1
    RecommendationRequest:
      type: object
      required:
        - answerOptionIds
      properties:
        answerOptionIds:
          type: array
          minItems: 1
          items:
            type: string
          example:
            - q1_a1
            - q2_a1
            - q3_a1
    RecommendationResponse:
      type: object
      required:
        - profile
        - items
        - totalItems
      properties:
        profile:
          $ref: "#/components/schemas/AromaProfile"
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationItem"
        totalItems:
          type: integer
          example: 4
    AromaProfile:
      type: object
      required:
        - name
        - description
        - tags
        - profileBars
        - characterTraits
        - keyNotes
      properties:
        name:
          type: string
          example: Спокойный минималист
        description:
          type: string
          example: Вам подходят чистые, спокойные и ненавязчивые ароматы.
        tags:
          type: array
          items:
            type: string
          example:
            - спокойный
            - чистый
            - офисный
        profileBars:
          type: array
          items:
            $ref: "#/components/schemas/ScoreMetric"
        characterTraits:
          type: array
          items:
            $ref: "#/components/schemas/ScoreMetric"
        keyNotes:
          type: array
          items:
            type: string
    ScoreMetric:
      type: object
      required:
        - label
        - percent
      properties:
        label:
          type: string
          example: Свежесть
        percent:
          type: integer
          minimum: 0
          maximum: 100
          example: 88
    RecommendationItem:
      type: object
      required:
        - id
        - name
        - brand
        - imageUrl
        - price
        - mainAccords
        - keyNotes
        - matchPercent
        - score
        - reason
      properties:
        id:
          type: string
          example: fresh-office
        name:
          type: string
          example: Fresh Office
        brand:
          type: string
          example: AromaType
        imageUrl:
          type: string
          example: ""
        price:
          type: string
          example: "490.00"
        mainAccords:
          type: array
          items:
            type: string
        keyNotes:
          type: array
          items:
            type: string
        matchPercent:
          type: integer
          example: 96
        score:
          type: integer
          example: 75
        reason:
          type: string
          example: "Совпадает с вашим профилем по тегам: свежий, чистый, офисный."
    Fragrance:
      type: object
      required:
        - id
        - name
        - brand
        - imageUrl
        - price
        - volumeOptions
        - description
        - topNotes
        - middleNotes
        - baseNotes
        - mainAccords
        - isActive
      properties:
        id:
          type: string
          example: fresh-office
        name:
          type: string
          example: Fresh Office
        brand:
          type: string
          example: AromaType
        imageUrl:
          type: string
          example: ""
        price:
          type: string
          example: "490.00"
        volumeOptions:
          type: array
          items:
            $ref: "#/components/schemas/VolumeOption"
        description:
          type: string
          example: Чистый и свежий аромат для учебы, офиса и спокойного повседневного образа.
        topNotes:
          type: array
          items:
            type: string
          example:
            - бергамот
            - лимон
        middleNotes:
          type: array
          items:
            type: string
          example:
            - лаванда
        baseNotes:
          type: array
          items:
            type: string
          example:
            - мускус
        mainAccords:
          type: array
          items:
            type: string
          example:
            - свежий
            - чистый
        isActive:
          type: boolean
          example: true
    CreateFragranceRequest:
      type: object
      required:
        - name
        - brand
      properties:
        name:
          type: string
          example: Citrus Day
        brand:
          type: string
          example: AromaType
        imageUrl:
          type: string
          example: https://example.com/citrus-day.jpg
        price:
          type: number
          example: 490
        volumeOptions:
          type: array
          items:
            $ref: "#/components/schemas/VolumeOption"
        description:
          type: string
          example: Легкий цитрусовый аромат на каждый день.
        topNotes:
          type: array
          items:
            type: string
          example:
            - лимон
            - бергамот
        middleNotes:
          type: array
          items:
            type: string
          example:
            - чай
        baseNotes:
          type: array
          items:
            type: string
          example:
            - мускус
        mainAccords:
          type: array
          items:
            type: string
          example:
            - свежий
        tagIds:
          type: array
          items:
            type: string
          example:
            - fresh
            - daily
        isActive:
          type: boolean
          default: true
    VolumeOption:
      type: object
      required:
        - volumeMl
        - price
      properties:
        volumeMl:
          type: integer
          example: 50
        price:
          type: number
          example: 490
    UploadResponse:
      type: object
      required:
        - imageUrl
      properties:
        imageUrl:
          type: string
          example: /uploads/0f4c1023-1d1d-4f67-a4d9-6d17ef6ff903.jpg